update register flow with create account button
